smart_monitoring-1/requirements.txt
flask
scikit-learn
numpy
pyjwt
paho-mqtt
redis
fastapi
uvicorn
tensorflow
pytest
python-dotenv